TTS-122: Cache UI for clear cache and Integration UI udpate

+++ api/TTA_Api_Routes.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace TTA_Api;
+
 use TTA\TTA_Cache;
 
 /**
@@ -332,12 +333,18 @@ class TTA_Api_Routes {
 		// save data about recording.
 		if ( 'post' == $request['method'] ) {
 			$fields = json_decode( $request['fields'] );
+			if ( isset( $fields->tta__settings_clear_all_cache ) && $fields->tta__settings_clear_all_cache ) {
+				TTA_Cache::flush();
+				$fields->tta__settings_clear_all_cache = false;
+			} else {
+				TTA_Cache::delete( 'all_settings' );
+			}
+
 
 			update_option( 'tta_settings_data', $fields );
 
 			$response['data'] = get_option( 'tta_settings_data' );
 
-			TTA_Cache::delete( 'all_settings' );
 
 			return rest_ensure_response( $response );
 		}
@@ -394,21 +401,21 @@ class TTA_Api_Routes {
 		}
 	}
 
-	public function get_all_user_roles($request) {
+	public function get_all_user_roles( $request ) {
 		// Access the global $wp_roles object
-		if (!isset($wp_roles)) {
+		if ( ! isset( $wp_roles ) ) {
 			global $wp_roles;
 		}
 
 		// Get all roles
 		$all_roles = $wp_roles->roles;
 
-		$user_roles = [];
+		$user_roles        = [];
 		$user_roles['all'] = 'All';
 
 		// Output all roles
-		foreach ($all_roles as $role_key => $role_data) {
-			$user_roles[$role_key]  = $role_data['name'];
+		foreach ( $all_roles as $role_key => $role_data ) {
+			$user_roles[ $role_key ] = $role_data['name'];
 		}
 
 		$response['status'] = true;
